+++ image/tools/github_pr_comment.py
@@ -5,6 +5,7 @@ import os
 import re
 import sys
 import datetime
+import hashlib
 from typing import Optional, Dict, Iterable
 
 import requests
@@ -98,25 +99,50 @@ def find_pr() -> str:
         raise Exception(f"The {event_type} event doesn\'t relate to a Pull Request.")
 
 def current_user() -> str:
+
+    token_hash = hashlib.sha256(os.environ["GITHUB_TOKEN"].encode()).hexdigest()
+
+    try:
+        with open(f'.dflook-terraform/token-cache/{token_hash}') as f:
+            username = f.read()
+            debug(f'GITHUB_TOKEN username: {username}')
+            return username
+    except Exception as e:
+        debug(str(e))
+
     response = github_api_request('get', 'https://api.github.com/user')
     if response.status_code != 403:
         user = response.json()
         debug('GITHUB_TOKEN user:')
         debug(json.dumps(user))
 
-        return user['login']
+        username = user['login']
+    else:
+        # Assume this is the github actions app token
+        username = 'github-actions[bot]'
 
-    # Assume this is the github actions app token
-    return 'github-actions[bot]'
+    try:
+        os.makedirs('.dflook-terraform/token-cache', exist_ok=True)
+        with open(f'.dflook-terraform/token-cache/{token_hash}', 'w') as f:
+            f.write(username)
+    except Exception as e:
+        debug(str(e))
+
+    debug(f'GITHUB_TOKEN username: {username}')
+    return username
 
 class TerraformComment:
     """
     The GitHub comment for this specific terraform plan
     """
 
-    def __init__(self, pr_url: str):
+    def __init__(self, pr_url: str=None):
         self._plan = None
         self._status = None
+        self._comment_url = None
+
+        if pr_url is None:
+            return
 
         response = github_api_request('get', pr_url)
         response.raise_for_status()
@@ -125,20 +151,20 @@ class TerraformComment:
         response = github_api_request('get', self._issue_url)
         response.raise_for_status()
 
-        self._comment_url = None
+        username = current_user()
+
         debug('Looking for an existing comment:')
         for comment in response.json():
             debug(json.dumps(comment))
-            if comment['user']['login'] == current_user():
-                match = re.match(rf'{re.escape(self._comment_identifier)}\n```(?:hcl)?(.*?)```(.*)', comment['body'], re.DOTALL)
+            if comment['user']['login'] == username:
+                match = re.match(rf'{re.escape(self._comment_identifier)}.*```(?:hcl)?(.*?)```.*', comment['body'], re.DOTALL)
 
                 if not match:
-                    match = re.match(rf'{re.escape(self._old_comment_identifier)}\n```(.*?)```(.*)', comment['body'], re.DOTALL)
+                    match = re.match(rf'{re.escape(self._old_comment_identifier)}\n```(.*?)```.*', comment['body'], re.DOTALL)
 
                 if match:
                     self._comment_url = comment['url']
                     self._plan = match.group(1).strip()
-                    self._status = match.group(2).strip()
                     return
 
     @property
@@ -270,12 +296,64 @@ class TerraformComment:
     def status(self, status: str) -> None:
         self._status = status.strip()
 
-    def update_comment(self):
+    def body(self) -> str:
         body = f'{self._comment_identifier}\n```hcl\n{self.plan}\n```'
 
         if self.status:
             body += '\n' + self.status
 
+        return body
+
+    def collapsable_body(self) -> str:
+
+        try:
+            collapse_threshold = int(os.environ['TF_PLAN_COLLAPSE_LENGTH'])
+        except (ValueError, KeyError):
+            collapse_threshold = 10
+
+        open = ''
+        highlighting = ''
+
+        if self.plan.startswith('Error'):
+            open = ' open'
+        elif 'Plan:' in self.plan:
+            highlighting = 'hcl'
+            num_lines = len(self.plan.splitlines())
+            if num_lines < collapse_threshold:
+                open = ' open'
+
+        body = f'''{self._comment_identifier}
+<details{open}>
+  <summary>{self.summary()}</summary>
+
+```{highlighting}
+{self.plan}
+```
+</details>
+'''
+
+        if self.status:
+            body += '\n' + self.status
+
+        return body
+
+    def summary(self) -> str:
+        summary = None
+
+        for line in self.plan.splitlines():
+            if line.startswith('No changes') or line.startswith('Error'):
+                return line
+
+            if line.startswith('Plan:'):
+                summary = line
+
+            if line.startswith('Changes to Outputs'):
+                return summary + ' Changes to Outputs.'
+
+        return summary
+
+    def update_comment(self):
+        body = self.collapsable_body()
         debug(body)
 
         if self._comment_url is None:
@@ -313,5 +391,6 @@ if __name__ == '__main__':
         if tf_comment.plan is None:
             exit(1)
         print(tf_comment.plan)
+        exit(0)
 
     tf_comment.update_comment()