+++ CHANGELOG.md
@@ -8,9 +8,20 @@ The actions are versioned as a suite. Some actions may have no change in behavio
 
 When using an action you can specify the version as:
 
-- `@v1.5.2` to use an exact release
-- `@v1.5` to use the latest patch release for the specific minor version
+- `@v1.6.0` to use an exact release
+- `@v1.6` to use the latest patch release for the specific minor version
 - `@v1` to use the latest patch release for the specific major version
+
+## [1.6.0] - 2021-02-25
+
+### Added
+- PR comments use a one line summary of the terraform output, with the full output in a collapsable pane.
+
+  If a plan is short the output is shown by default. This can be controlled with the `TF_PLAN_COLLAPSE_LENGTH` environment
+  variable for the [dflook/terraform-plan](terraform-plan) action.
+
+### Fixed
+- Now makes far fewer github api requests to avoid rate limiting.
 
 ## [1.5.2] - 2021-01-16
 
@@ -95,6 +106,7 @@ First release of the GitHub Actions:
 - [dflook/terraform-new-workspace](terraform-new-workspace)
 - [dflook/terraform-destroy-workspace](terraform-destroy-workspace)
 
+[1.6.0]: https://github.com/dflook/terraform-github-actions/compare/v1.5.2...v1.6.0
 [1.5.2]: https://github.com/dflook/terraform-github-actions/compare/v1.5.1...v1.5.2
 [1.5.1]: https://github.com/dflook/terraform-github-actions/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/dflook/terraform-github-actions/compare/v1.4.2...v1.5.0